# autopilot-api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/content-ai/generate": {
      "post": {
        "operationId": "ContentAiController_generate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateContentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Content AI"
        ]
      }
    },
    "/api/v1/content-ai/repurpose": {
      "post": {
        "operationId": "ContentAiController_repurpose",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepurposeContentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Content AI"
        ]
      }
    },
    "/api/v1/content-ai/adapt-platforms": {
      "post": {
        "operationId": "ContentAiController_adapt",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdaptPlatformsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Content AI"
        ]
      }
    },
    "/api/v1/content-ai/generate-image": {
      "post": {
        "operationId": "ContentAiController_image",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateImageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Content AI"
        ]
      }
    },
    "/api/v1/content-ai/generate-slideshow": {
      "post": {
        "operationId": "ContentAiController_slideshow",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateSlideshowDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Content AI"
        ]
      }
    },
    "/api/v1/content-ai/auto-publish": {
      "post": {
        "operationId": "ContentAiController_runAutoPublish",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Content AI"
        ]
      }
    },
    "/api/v1/content-ai/daily-workflow": {
      "post": {
        "operationId": "ContentAiController_runDailyWorkflow",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DailyWorkflowDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Content AI"
        ]
      }
    },
    "/api/v1/content-ai/{contentId}/publish": {
      "post": {
        "operationId": "ContentAiController_publish",
        "parameters": [
          {
            "name": "contentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishContentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Content AI"
        ]
      }
    },
...
    "/api/v1/comment-replies/fetch": {
      "post": {
        "operationId": "CommentRepliesController_fetch",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FetchCommentsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Comment Replies"
        ]
      }
    },
    "/api/v1/comment-replies/{id}/suggest": {
      "post": {
        "operationId": "CommentRepliesController_suggest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Comment Replies"
        ]
      }
    },
    "/api/v1/comment-replies/{id}/send": {
      "post": {
        "operationId": "CommentRepliesController_send",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendCommentReplyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Comment Replies"
        ]
      }
    },
    "/api/v1/comment-replies": {
      "post": {
        "operationId": "CommentRepliesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentRepliesCreateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Comment Replies"
        ]
      },
      "get": {
        "operationId": "CommentRepliesController_findAll",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Comment Replies"
        ]
      }
    },
    "/api/v1/comment-replies/inbox": {
      "get": {
        "operationId": "CommentRepliesController_getInbox",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contentId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Comment Replies"
        ]
      }
    },
    "/api/v1/comment-replies/{id}": {
      "get": {
        "operationId": "CommentRepliesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Comment Replies"
        ]
      },
      "patch": {
        "operationId": "CommentRepliesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentRepliesUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Comment Replies"
        ]
      },
      "delete": {
        "operationId": "CommentRepliesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Comment Replies"
        ]
      }
    },
...
  },
  "info": {
    "title": "API",
...
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "widget-api-key": {
        "type": "http",
        "in": "header",
        "name": "widget-api-key",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "Chatbot widget API key (`pk_live_…`). Create keys in AI Chatbot → Embed. Used for `/api/v1/widget/*` routes."
      }
    },
    "schemas": {
...
      "GenerateContentDto": {
        "type": "object",
        "properties": {}
      },
      "RepurposeContentDto": {
        "type": "object",
        "properties": {}
      },
      "AdaptPlatformsDto": {
        "type": "object",
        "properties": {}
      },
      "GenerateImageDto": {
        "type": "object",
        "properties": {}
      },
      "GenerateSlideshowDto": {
        "type": "object",
        "properties": {}
      },
      "DailyWorkflowDto": {
        "type": "object",
        "properties": {}
      },
      "PublishContentDto": {
        "type": "object",
        "properties": {}
      },
...
      "FetchCommentsDto": {
        "type": "object",
        "properties": {}
      },
      "SendCommentReplyDto": {
        "type": "object",
        "properties": {}
      },
      "CommentRepliesCreateDto": {
        "type": "object",
        "properties": {}
      },
      "CommentRepliesUpdateDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}